random voice-led piano hexachords
Chooses a four note baseline and hexachords.
Voices them with some basic voice-leading.

Switched to piano output instead of my simple synth instrument.
This renders to Haskore, to a tmp file, to timidity.
This is fine for now. For arbitrarily modified audio we have to
complete the code working with the Soundfont standard.